functions/requirements.txt
azure-common==1.1.28
azure-core==1.25.1
azure-functions==1.11.2
azure-identity==1.10.0
azure-mgmt-core==1.3.2
azure-mgmt-deploymentmanager==1.0.0
azure-mgmt-resource==21.1.0
PyNaCl==1.5.0